+++ MekHQ/src/mekhq/campaign/mission/newContract/ContractXmlCodec.java
@@ -57,6 +57,7 @@ import mekhq.campaign.mission.enums.MissionStatus;
 import mekhq.campaign.mission.newContract.contractData.*;
 import mekhq.campaign.mission.newContract.contractGeneration.ChaosEmployerType;
 import mekhq.campaign.mission.newContract.contractGeneration.ContractTermsData;
+import mekhq.campaign.mission.newContract.contractGeneration.negotiationsAndNPCs.TermFunding;
 import mekhq.campaign.personnel.Person;
 import mekhq.utilities.MHQXMLUtility;
 import org.w3c.dom.NamedNodeMap;
