+++ MekHQ/src/mekhq/gui/dialog/markets/contractMarket/ContractNegotiationDialog.java
@@ -71,8 +71,8 @@ import mekhq.campaign.mission.newContract.AbstractContract;
 import mekhq.campaign.mission.newContract.contractData.ChaosContractStepsTable;
 import mekhq.campaign.mission.newContract.contractData.NegotiationData;
 import mekhq.campaign.mission.newContract.contractData.RentedFacilitiesData;
-import mekhq.campaign.mission.newContract.contractData.TermFunding;
 import mekhq.campaign.mission.newContract.contractGeneration.ContractTermsData;
+import mekhq.campaign.mission.newContract.contractGeneration.negotiationsAndNPCs.TermFunding;
 import mekhq.campaign.mission.newContract.contractGeneration.targetFinder.ChaosContractPayDetermination;
 import mekhq.campaign.personnel.Person;
 import mekhq.gui.baseComponents.roundedComponents.RoundedJButton;
